estimation1/requirements.txt
numpy==1.19.5
Pillow==9.5.0
pyopencl==2022.1
scikit_learn==1.2.2
tensorflow==2.12.0
tensorflow_cpu==2.11.0
